Took out old flag communication and added in new one
The new way is using a bitmask as well as the size of the map to our advantage. This uses the MOD 128 (or as a bitmask) to move the x & y values for the Enemy Enlightenment Center.

I also took out the old communication flags. Added more to the enlightenmentcenter code.