Fixed Bugs, Added Pathfinding, Made Smarter Poli EC
The biggest thing with this was I wanted to make pathing actually work. I ended up using a small greedy algorithm and it is way more improvement than not doing anything. The guys are getting to places a LOT faster.
Another thing was taking out some code that made EC bombs go off even when they knew where locations where to go to.
*Defenders should now spread out instead of bumching up together as much.
*Bug with code not building slanderers if we were stocking up for Enemy EC